[DEV-6386] Move tests for calc_metrics_deployments to their module
