+++ server/tests/controllers/test_metrics_controller.py
@@ -2,18 +2,14 @@ from datetime import date
 import json
 
 import pytest
-from sqlalchemy import delete, insert
 
 from athenian.api.internal.miners.github import developer
-from athenian.api.models.persistentdata.models import DeployedComponent, DeployedLabel
 from athenian.api.models.web import (
     CalculatedCodeCheckMetrics,
-    CalculatedDeploymentMetric,
     CalculatedDeveloperMetrics,
     CalculatedLinearMetricValues,
     CodeBypassingPRsMeasurement,
     CodeCheckMetricID,
-    DeploymentMetricID,
     DeveloperMetricID,
 )
 from athenian.api.serialization import FriendlyJson
@@ -1185,464 +1181,6 @@ async def test_empty_granularities(client, headers):
     }
     response = await client.request(
         method="POST", path="/v1/metrics/code_checks", headers=headers, json=body,
-    )
-    assert response.status == 400
-    res = await response.json()
-    assert "granularities" in res["detail"]
-
-
-# TODO: fix response validation against the schema
-@pytest.mark.app_validate_responses(False)
-async def test_deployment_metrics_smoke(client, headers, sample_deployments):
-    body = {
-        "account": 1,
-        "date_from": "2018-01-12",
-        "date_to": "2020-03-01",
-        "for": [
-            {
-                "repositories": ["{1}"],
-                "withgroups": [
-                    {"releaser": ["github.com/mcuadros"]},
-                    {"pr_author": ["github.com/mcuadros"]},
-                ],
-                "environments": ["staging", "production", "mirror"],
-            },
-        ],
-        "metrics": [
-            DeploymentMetricID.DEP_SUCCESS_COUNT,
-            DeploymentMetricID.DEP_DURATION_SUCCESSFUL,
-        ],
-        "granularities": ["all"],
-    }
-    response = await client.request(
-        method="POST", path="/v1/metrics/deployments", headers=headers, json=body,
-    )
-    body = (await response.read()).decode("utf-8")
-    assert response.status == 200, "Response body is : " + body
-    model = [CalculatedDeploymentMetric.from_dict(obj) for obj in json.loads(body)]
-    assert [m.to_dict() for m in model] == [
-        {
-            "for": {
-                "repositories": ["{1}"],
-                "with": {"releaser": ["github.com/mcuadros"]},
-                "environments": ["staging"],
-            },
-            "metrics": ["dep-success-count", "dep-duration-successful"],
-            "granularity": "all",
-            "values": [
-                {
-                    "date": date(2018, 1, 12),
-                    "values": [3, "600s"],
-                    "confidence_maxs": [None, "600s"],
-                    "confidence_mins": [None, "600s"],
-                    "confidence_scores": [None, 100],
-                },
-            ],
-        },
-        {
-            "for": {
-                "repositories": ["{1}"],
-                "with": {"releaser": ["github.com/mcuadros"]},
-                "environments": ["production"],
-            },
-            "metrics": ["dep-success-count", "dep-duration-successful"],
-            "granularity": "all",
-            "values": [
-                {
-                    "date": date(2018, 1, 12),
-                    "values": [3, "600s"],
-                    "confidence_maxs": [None, "600s"],
-                    "confidence_mins": [None, "600s"],
-                    "confidence_scores": [None, 100],
-                },
-            ],
-        },
-        {
-            "for": {
-                "repositories": ["{1}"],
-                "with": {"releaser": ["github.com/mcuadros"]},
-                "environments": ["mirror"],
-            },
-            "metrics": ["dep-success-count", "dep-duration-successful"],
-            "granularity": "all",
-            "values": [
-                {
-                    "date": date(2018, 1, 12),
-                    "values": [0, None],
-                },
-            ],
-        },
-        {
-            "for": {
-                "repositories": ["{1}"],
-                "with": {"pr_author": ["github.com/mcuadros"]},
-                "environments": ["staging"],
-            },
-            "metrics": ["dep-success-count", "dep-duration-successful"],
-            "granularity": "all",
-            "values": [
-                {
-                    "date": date(2018, 1, 12),
-                    "values": [3, "600s"],
-                    "confidence_maxs": [None, "600s"],
-                    "confidence_mins": [None, "600s"],
-                    "confidence_scores": [None, 100],
-                },
-            ],
-        },
-        {
-            "for": {
-                "repositories": ["{1}"],
-                "with": {"pr_author": ["github.com/mcuadros"]},
-                "environments": ["production"],
-            },
-            "metrics": ["dep-success-count", "dep-duration-successful"],
-            "granularity": "all",
-            "values": [
-                {
-                    "date": date(2018, 1, 12),
-                    "values": [3, "600s"],
-                    "confidence_maxs": [None, "600s"],
-                    "confidence_mins": [None, "600s"],
-                    "confidence_scores": [None, 100],
-                },
-            ],
-        },
-        {
-            "for": {
-                "repositories": ["{1}"],
-                "with": {"pr_author": ["github.com/mcuadros"]},
-                "environments": ["mirror"],
-            },
-            "metrics": ["dep-success-count", "dep-duration-successful"],
-            "granularity": "all",
-            "values": [
-                {
-                    "date": date(2018, 1, 12),
-                    "values": [0, None],
-                },
-            ],
-        },
-    ]
-
-
-async def test_deployment_metrics_empty_for(client, headers, sample_deployments):
-    body = {
-        "account": 1,
-        "date_from": "2018-01-12",
-        "date_to": "2020-03-01",
-        "for": [{}],
-        "metrics": [DeploymentMetricID.DEP_SUCCESS_COUNT],
-        "granularities": ["all"],
-    }
-    response = await client.request(
-        method="POST", path="/v1/metrics/deployments", headers=headers, json=body,
-    )
-    body = (await response.read()).decode("utf-8")
-    assert response.status == 200, "Response body is : " + body
-    model = [CalculatedDeploymentMetric.from_dict(obj) for obj in json.loads(body)]
-    assert [m.to_dict() for m in model] == [
-        {
-            "for": {},
-            "granularity": "all",
-            "metrics": [DeploymentMetricID.DEP_SUCCESS_COUNT],
-            "values": [{"date": date(2018, 1, 12), "values": [9]}],
-        },
-    ]
-
-
-@pytest.mark.parametrize(
-    "account, date_from, date_to, repos, withgroups, metrics, code",
-    [
-        (1, "2018-01-12", "2020-01-12", ["{1}"], [], [DeploymentMetricID.DEP_PRS_COUNT], 200),
-        (1, "2020-01-12", "2018-01-12", ["{1}"], [], [DeploymentMetricID.DEP_PRS_COUNT], 400),
-        (
-            2,
-            "2018-01-12",
-            "2020-01-12",
-            ["github.com/src-d/go-git"],
-            [],
-            [DeploymentMetricID.DEP_PRS_COUNT],
-            422,
-        ),
-        (
-            3,
-            "2018-01-12",
-            "2020-01-12",
-            ["github.com/src-d/go-git"],
-            [],
-            [DeploymentMetricID.DEP_PRS_COUNT],
-            404,
-        ),
-        (1, "2018-01-12", "2020-01-12", ["{1}"], [], ["whatever"], 400),
-        (
-            1,
-            "2018-01-12",
-            "2020-01-12",
-            ["github.com/athenianco/athenian-api"],
-            [],
-            [DeploymentMetricID.DEP_PRS_COUNT],
-            403,
-        ),
-        (
-            1,
-            "2018-01-12",
-            "2020-01-12",
-            ["{1}"],
-            [{"pr_author": ["github.com/akbarik"]}],
-            [DeploymentMetricID.DEP_PRS_COUNT],
-            400,
-        ),
-    ],
-)
-async def test_deployment_metrics_nasty_input(
-    client,
-    headers,
-    account,
-    date_from,
-    date_to,
-    repos,
-    withgroups,
-    metrics,
-    code,
-):
-    body = {
-        "account": account,
-        "date_from": date_from,
-        "date_to": date_to,
-        "for": [
-            {
-                "repositories": [*repos],
-                "withgroups": [*withgroups],
-            },
-        ],
-        "metrics": [*metrics],
-        "granularities": ["all"],
-    }
-    response = await client.request(
-        method="POST", path="/v1/metrics/deployments", headers=headers, json=body,
-    )
-    body = (await response.read()).decode("utf-8")
-    assert response.status == code, "Response body is : " + body
-
-
-async def test_deployment_metrics_filter_labels(
-    client,
-    headers,
-    precomputed_deployments,
-    rdb,
-    client_cache,
-):
-    body = {
-        "account": 1,
-        "date_from": "2015-01-12",
-        "date_to": "2020-03-01",
-        "for": [
-            {
-                "pr_labels_include": ["bug", "plumbing", "enhancement"],
-            },
-        ],
-        "metrics": [DeploymentMetricID.DEP_COUNT],
-        "granularities": ["all"],
-    }
-
-    async def request():
-        response = await client.request(
-            method="POST", path="/v1/metrics/deployments", headers=headers, json=body,
-        )
-        rbody = (await response.read()).decode("utf-8")
-        assert response.status == 200, "Response body is : " + rbody
-        return [CalculatedDeploymentMetric.from_dict(obj) for obj in json.loads(rbody)]
-
-    model = await request()
-    assert model[0].values[0].values[0] == 1
-
-    del body["for"][0]["pr_labels_include"]
-    body["for"][0]["pr_labels_exclude"] = ["bug", "plumbing", "enhancement"]
-    model = await request()
-    assert model[0].values[0].values[0] == 0
-
-    del body["for"][0]["pr_labels_exclude"]
-    await rdb.execute(
-        insert(DeployedLabel).values(
-            {
-                DeployedLabel.account_id: 1,
-                DeployedLabel.deployment_name: "Dummy deployment",
-                DeployedLabel.key: "nginx",
-                DeployedLabel.value: 504,
-            },
-        ),
-    )
-    body["for"][0]["with_labels"] = {"nginx": 503}
-    model = await request()
-    assert model[0].values[0].values[0] == 0
-
-    body["for"][0]["with_labels"] = {"nginx": 504}
-    model = await request()
-    assert model[0].values[0].values[0] == 1
-
-    del body["for"][0]["with_labels"]
-    body["for"][0]["without_labels"] = {"nginx": 503}
-    model = await request()
-    assert model[0].values[0].values[0] == 1
-
-    body["for"][0]["without_labels"] = {"nginx": 504}
-    model = await request()
-    assert model[0].values[0].values[0] == 0
-
-
-async def test_deployment_metrics_environments(
-    client,
-    headers,
-    sample_deployments,
-    rdb,
-    client_cache,
-):
-    await rdb.execute(
-        delete(DeployedComponent).where(DeployedComponent.deployment_name == "staging_2018_12_02"),
-    )
-    body = {
-        "account": 1,
-        "date_from": "2018-01-12",
-        "date_to": "2020-03-01",
-        "for": [
-            {
-                "envgroups": [["production"]],
-            },
-        ],
-        "metrics": [
-            DeploymentMetricID.DEP_SUCCESS_COUNT,
-            DeploymentMetricID.DEP_DURATION_SUCCESSFUL,
-        ],
-        "granularities": ["all"],
-    }
-
-    async def request():
-        response = await client.request(
-            method="POST", path="/v1/metrics/deployments", headers=headers, json=body,
-        )
-        rbody = (await response.read()).decode("utf-8")
-        assert response.status == 200, "Response body is : " + rbody
-        return [CalculatedDeploymentMetric.from_dict(obj) for obj in json.loads(rbody)]
-
-    model = await request()
-    assert len(model) == 1
-    assert model[0].values[0].values[0] == 3
-    body["for"][0]["envgroups"] = [["staging"], ["production"]]
-    model = await request()
-    assert len(model) == 2
-    assert model[0].for_.environments == ["staging"]
-    assert model[0].values[0].values[0] == 3
-    assert model[1].for_.environments == ["production"]
-    assert model[1].values[0].values[0] == 3
-
-
-async def test_deployment_metrics_with(client, headers, sample_deployments):
-    body = {
-        "account": 1,
-        "date_from": "2018-01-12",
-        "date_to": "2020-03-01",
-        "for": [
-            {
-                "with": {"pr_author": ["github.com/mcuadros"]},
-                "environments": ["production"],
-            },
-        ],
-        "metrics": [
-            DeploymentMetricID.DEP_SUCCESS_COUNT,
-            DeploymentMetricID.DEP_DURATION_SUCCESSFUL,
-        ],
-        "granularities": ["all"],
-    }
-
-    response = await client.request(
-        method="POST", path="/v1/metrics/deployments", headers=headers, json=body,
-    )
-    rbody = (await response.read()).decode("utf-8")
-    assert response.status == 200, "Response body is : " + rbody
-    model = [CalculatedDeploymentMetric.from_dict(obj) for obj in json.loads(rbody)]
-
-    assert len(model) == 1
-    assert model[0].values[0].values[0] == 3
-
-
-@pytest.mark.app_validate_responses(False)
-async def test_deployment_metrics_team(client, headers, sample_deployments, sample_team):
-    team_str = "{%d}" % sample_team
-    body = {
-        "account": 1,
-        "date_from": "2018-01-12",
-        "date_to": "2020-03-01",
-        "for": [
-            {
-                "repositories": ["{1}"],
-                "withgroups": [{"releaser": [team_str]}, {"pr_author": [team_str]}],
-                "environments": ["production"],
-            },
-        ],
-        "metrics": [
-            DeploymentMetricID.DEP_SUCCESS_COUNT,
-            DeploymentMetricID.DEP_DURATION_SUCCESSFUL,
-        ],
-        "granularities": ["all"],
-    }
-    response = await client.request(
-        method="POST", path="/v1/metrics/deployments", headers=headers, json=body,
-    )
-    body = (await response.read()).decode("utf-8")
-    assert response.status == 200, "Response body is : " + body
-    model = [CalculatedDeploymentMetric.from_dict(obj) for obj in json.loads(body)]
-    assert [m.to_dict() for m in model] == [
-        {
-            "for": {
-                "repositories": ["{1}"],
-                "with": {"releaser": [team_str]},
-                "environments": ["production"],
-            },
-            "metrics": ["dep-success-count", "dep-duration-successful"],
-            "granularity": "all",
-            "values": [
-                {
-                    "date": date(2018, 1, 12),
-                    "values": [3, "600s"],
-                    "confidence_maxs": [None, "600s"],
-                    "confidence_mins": [None, "600s"],
-                    "confidence_scores": [None, 100],
-                },
-            ],
-        },
-        {
-            "for": {
-                "repositories": ["{1}"],
-                "with": {"pr_author": [team_str]},
-                "environments": ["production"],
-            },
-            "metrics": ["dep-success-count", "dep-duration-successful"],
-            "granularity": "all",
-            "values": [
-                {
-                    "date": date(2018, 1, 12),
-                    "values": [3, "600s"],
-                    "confidence_maxs": [None, "600s"],
-                    "confidence_mins": [None, "600s"],
-                    "confidence_scores": [None, 100],
-                },
-            ],
-        },
-    ]
-
-
-async def test_deployment_metrics_empty_granularities(client, headers, sample_deployments):
-    body = {
-        "account": 1,
-        "date_from": "2018-01-12",
-        "date_to": "2020-03-01",
-        "for": [{}],
-        "metrics": [DeploymentMetricID.DEP_SUCCESS_COUNT],
-        "granularities": [],
-    }
-    response = await client.request(
-        method="POST", path="/v1/metrics/deployments", headers=headers, json=body,
     )
     assert response.status == 400
     res = await response.json()
